chore: storybook preview에 decorators 추가 (#11)
